More renaming
Rename some of the files so that they all follow some form of
convention:
- files that contain functions registered as typer app commends (have
  @<app>.command) need to end in `_cmd` or `commands` for clarity
- files that define a new typer app need to have `app` somewhere in
  their name
- deployer commands that don't require defining any new helper files
  (like cilogon) don't need to be into their own directory by
  themselves. Instead keep them in the parent `deployer/commands`, under
  a name follwing the two rules above
