+++ deployer/commands/generate/dedicated_cluster/aws_commands.py
@@ -14,8 +14,8 @@ import typer
 
 from deployer.utils.rendering import print_colour
 
-from .app import dedicated_cluster_app
 from .common import REPO_ROOT, generate_config_directory, generate_support_files
+from .dedicate_cluster_app import dedicated_cluster_app
 
 
 def generate_infra_files(vars):
