+++ deployer/commands/generate/dedicated_cluster/dedicate_cluster_app.py
