+++ deployer/commands/generate/dedicated_cluster/gcp_commands.py
@@ -15,13 +15,13 @@ from typing_extensions import Annotated
 
 from deployer.utils.rendering import print_colour
 
-from .app import dedicated_cluster_app
 from .common import (
     REPO_ROOT,
     generate_cluster_config_file,
     generate_config_directory,
     generate_support_files,
 )
+from .dedicate_cluster_app import dedicated_cluster_app
 
 
 def generate_terraform_file(vars):
