+++ deployer/commands/validate/config_cmd.py
